backend/requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector DB
numpy<2.0
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0

# LLM & AI
google-generativeai==0.3.2

# Firebase
firebase-admin==6.3.0
python-firebase==1.2

# Database & Storage
sqlalchemy==2.0.23
pymongo==4.6.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1

# Email Processing
beautifulsoup4==4.12.2
mail-parser
lxml